Type/brand forms

+++ server/src/services/typeService.ts
@@ -2,7 +2,7 @@ import prisma from '../config/prismaClient';
 import ApiError from '../error/ApiError';
 import ErrorMessage from '../error/errorMessage';
 
-class ProductTypeService {
+class TypeService {
   static async createType(name: string) {
     const foundType = await prisma.productType.findFirst({ where: { name } });
     if (foundType) throw new ApiError(409, ErrorMessage.PRODUCT_TYPE_EXISTS);
@@ -15,20 +15,15 @@ class ProductTypeService {
     return types;
   }
 
-  static async updateType(id: string, name: string) {
-    const type = await prisma.productType
-      .update({ where: { id }, data: { name } })
-      .catch(() => {
-        throw new ApiError(404, ErrorMessage.PRODUCT_TYPE_NOT_FOUND);
-      });
-    return type;
-  }
-
   static async deleteType(id: string) {
+    const hasProducts = await prisma.product.findFirst({
+      where: { typeId: id },
+    });
+    if (hasProducts) throw new ApiError(400, ErrorMessage.BAD_REQUEST);
     await prisma.productType.delete({ where: { id } }).catch(() => {
       throw new ApiError(404, ErrorMessage.PRODUCT_TYPE_NOT_FOUND);
     });
   }
 }
 
-export default ProductTypeService;
+export default TypeService;
